Move `generate` CLI command logic to ExTTRPGDev.CLI.Generate
